\name{ismdi}
\alias{ismdi}
\title{Check Window mode (sdi, mdi) of RGui}
\description{
    This functions reports whether RGui was started in --mdi or --sdi mode.
}
\usage{
ismdi()
}
\value{  
    \item{mdi}{Returns \code{TRUE}, if RGui was started in its default --mdi mode, else \code{FALSE}.}
}
\keyword{misc}
